+++ DSCResources/MSFT_SqlSetup/MSFT_SqlSetup.psm1
@@ -1574,7 +1574,16 @@ function Set-TargetResource
                 }
                 else
                 {
-                    $setupArgumentValue = '"{0}"' -f $currentSetupArgument.Value
+                    # Logic added as a fix for Issue#1254 SqlSetup:Fails when a root directory is specified
+                    if($currentSetupArgument.Value -match '^[a-zA-Z]:\\$')
+                    {
+                        $setupArgumentValue = $currentSetupArgument.Value   
+                    }
+                    else 
+                    {
+                        $setupArgumentValue = '"{0}"' -f $currentSetupArgument.Value  
+                    }
+                    
                 }
             }